move ComputeFeeDistribution into store, next to the type it returns
Review feedback on #917 (tamirms): the percentile math was copied
verbatim into v2's feewindow to avoid sharing v1's baggage — but this
one function is pure ([]uint64, uint32) -> store.FeeDistribution over
stdlib only, so the baggage argument never applied to it. It now lives
in internal/store beside FeeDistribution, both services' fee windows
call it, and the differential test against the stats library exists
once instead of twice. This is the one piece where silent drift would
actually matter: both services' getFeeStats numbers come out of this
function. The windows, classification, and retention machinery stay
per-tree.
